% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_GEFS_Sf_ensembles.R
\name{cal_GEFS_Sf_ensembles}
\alias{cal_GEFS_Sf_ensembles}
\title{Loads the GEFS weather forecast data ensembles}
\usage{
cal_GEFS_Sf_ensembles(fileNames)
}
\arguments{
\item{fileNames}{A list of the GEFS filename}

\item{dates}{The list of dates for the forecast}
}
\description{
Loads the GEFS weather forecast data ensembles
}
