% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSf.R
\name{calSf}
\alias{calSf}
\title{Calculates cummulative Tair}
\usage{
calSf(Tairs, days)
}
\arguments{
\item{Tairs}{The start date of desired data download}

\item{days}{The desired dates}
}
\description{
Calculates cummulative Tair
}
