% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calStationScore.R
\name{calStationScore}
\alias{calStationScore}
\title{Calculates a met station score based on Hopkin's Bioclimatic Law}
\usage{
calStationScore(siteLat, siteLong, siteElev, statLat, statLong, statElev)
}
\arguments{
\item{statElev}{}
}
\value{

}
\description{
Calculates a met station score based on Hopkin's Bioclimatic Law
}
