% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInits_forecast.R
\name{createInits_forecast}
\alias{createInits_forecast}
\title{Create inits based on previous out.burn}
\usage{
createInits_forecast(out.burn, variableNames)
}
\arguments{
\item{variableNames}{}
}
\description{
Create inits based on previous out.burn
}
