% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSf.R
\name{createSf}
\alias{createSf}
\title{Creates data object with the desired Sf means and variances}
\usage{
createSf(lat = "", long = "", years, siteName, dataDirectory, endDate,
  GEFS_Files, GEFS_Directory)
}
\arguments{
\item{siteName}{The site name}

\item{dataDirectory}{The data directory}

\item{endDate}{The end date}

\item{GEFS_Files}{A vector of the file names for the GEFS files}

\item{GEFS_Directory}{The directory where the GEFS files are located}

\item{latitude}{The desired latitude if not willowCreek}

\item{longitude}{The desired longitude if not willowCreek}
}
\description{
Creates data object with the desired Sf means and variances
}
