% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.phenocam.R
\name{download.phenocam}
\alias{download.phenocam}
\title{Download Phenocam data}
\usage{
download.phenocam(URL)
}
\arguments{
\item{URL}{web address where data is located}
}
\description{
Download Phenocam data
}
