% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_US_WCr_met.R
\name{download_US_WCr_met}
\alias{download_US_WCr_met}
\title{Downloads the meteological data for Willow Creek}
\usage{
download_US_WCr_met(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of desired data download}

\item{end_date}{The end date of desired data download}
}
\description{
Downloads the meteological data for Willow Creek
}
