% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMODIS.R
\name{downloadMODIS}
\alias{downloadMODIS}
\title{Function to download all MODIS data for input into forecast model}
\usage{
downloadMODIS(startDate, endDate, dataDirectory, lat, long, siteName)
}
\arguments{
\item{startDate}{The start date of data in date format}

\item{endDate}{The end date of data in date format}

\item{dataDirectory}{The file path for the directory to download and store data in}

\item{lat}{Latitide of the site in decimal degrees}

\item{long}{Longitude of the site in decimal degrees}

\item{siteName}{The site name for file naming}
}
\description{
Function to download all MODIS data for input into forecast model
}
