% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNOAAlag.R
\name{fillNOAAlag}
\alias{fillNOAAlag}
\title{Fills the missing data due to lag in NOAA met with GEFS forecasts}
\usage{
fillNOAAlag(days, siteName)
}
\arguments{
\item{fileName}{The GEFS filename}

\item{dataDirectory}{The directory that the data is stored in}
}
\description{
Fills the missing data due to lag in NOAA met with GEFS forecasts
}
