% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastLog.R
\name{forecastLog}
\alias{forecastLog}
\title{Basic logistic forecast step}
\usage{
forecastLog(IC, r, Q = 0, n, NT)
}
\arguments{
\item{IC}{Initial conditions}

\item{r}{The parameter r}

\item{Q}{Process error (default = 0 for deterministic runs)}

\item{n}{Size of Monte Carlo ensemble}

\item{NT}{number of days for forecast}
}
\value{

}
\description{
Basic logistic forecast step
}
