% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastLogCov2.R
\name{forecastLogCov2}
\alias{forecastLogCov2}
\title{Logistic with covariates forecast step}
\usage{
forecastLogCov2(IC, trans, b1, Sf, Q = 0, n, NT)
}
\arguments{
\item{IC}{Initial conditions}

\item{trans}{The parameter trans}

\item{b1}{The parameter b1}

\item{Sf}{The covariate Sf (cummulative average temperatures)}

\item{Q}{Process error (default = 0 for deterministic runs)}

\item{n}{Size of Monte Carlo ensemble}

\item{NT}{number of days for forecast}
}
\value{

}
\description{
Logistic with covariates forecast step
}
