% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastRW.R
\name{forecastRW}
\alias{forecastRW}
\title{Random walk forecast step}
\usage{
forecastRW(IC, Q = 0, n, NT)
}
\arguments{
\item{IC}{Initial conditions}

\item{Q}{Process error (default = 0 for deterministic runs)}

\item{n}{Size of Monte Carlo ensemble}

\item{NT}{number of days for forecast}
}
\value{

}
\description{
Random walk forecast step
}
