% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationData.R
\name{getStationData}
\alias{getStationData}
\title{Gets data for a specific station with a tryCatch}
\usage{
getStationData(station, startDate, endDate)
}
\value{

}
\description{
Gets data for a specific station with a tryCatch
}
