% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNOAAstation.R
\name{identifyNOAAstation}
\alias{identifyNOAAstation}
\title{Identifies the best NOAA station}
\usage{
identifyNOAAstation(lat, long, elev, siteName, startDate, endDate)
}
\arguments{
\item{lat}{Site latitude}

\item{long}{Site longitude}

\item{elev}{Site elevation (m)}

\item{startDate}{Start date}

\item{endDate}{End Date}
}
\description{
Identifies the best NOAA station
}
