% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ERA5_Tair.R
\name{load_ERA5_Tair}
\alias{load_ERA5_Tair}
\title{Calculates cummulative Tair (within one year for one ensemble)}
\usage{
load_ERA5_Tair(lat, long, years)
}
\arguments{
\item{lat}{The site latitude}

\item{long}{The site longitude}

\item{years}{The desired years to download}
}
\description{
Calculates cummulative Tair (within one year for one ensemble)
}
