% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NOAA_met.R
\name{load_NOAA_met}
\alias{load_NOAA_met}
\title{Calculates cummulative Tair (within one year for one ensemble)}
\usage{
load_NOAA_met(lat, long, years)
}
\arguments{
\item{lat}{The site latitude}

\item{long}{The site longitude}

\item{years}{The desired years}
}
\description{
Calculates cummulative Tair (within one year for one ensemble)
}
