% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoModel_CDD_Autumn_GCC.R
\name{phenoModel_CDD_Autumn_GCC}
\alias{phenoModel_CDD_Autumn_GCC}
\title{Creates a logistic phenology forecast model based on PhenoCam}
\usage{
phenoModel_CDD_Autumn_GCC(data, nchain, MODIS_index = "NDVI", baseTemp = NA)
}
\arguments{
\item{data}{The data in the form of a list with data$p, data$mn, data$me, data$n, data$x_ic, and data$tau_ic}

\item{nchain}{The desired number of chains in the MCMC}

\item{baseTemp}{The desired base temperature}
}
\description{
Creates a logistic phenology forecast model based on PhenoCam
}
