% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoModel_CDD_Autumn_NDVI.R
\name{phenoModel_CDD_Autumn_NDVI}
\alias{phenoModel_CDD_Autumn_NDVI}
\title{Creates a logistic phenology forecast model based on PhenoCam and MODIS data}
\usage{
phenoModel_CDD_Autumn_NDVI(data, nchain, baseTemp = NA)
}
\arguments{
\item{data}{The data in the form of a list with data$p, data$mn, data$me, data$n, data$x_ic, and data$tau_ic}

\item{nchain}{The desired number of chains in the MCMC}

\item{baseTemp}{The base temperature for CDD calculations}
}
\description{
Creates a logistic phenology forecast model based on PhenoCam and MODIS data
}
