% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMODIS.R
\name{prepareMODIS}
\alias{prepareMODIS}
\title{Function to prepare MODIS data for input into forecast model}
\usage{
prepareMODIS(startDate, endDate, metric, timeForecast, dataDirectory,
  lat = "", long = "", siteName)
}
\arguments{
\item{startDate}{The start date of data in date format}

\item{endDate}{The end date of data in date format}

\item{metric}{NDVI or EVI}

\item{timeForecast}{The time values (in days) based in PhenoCam available data}

\item{dataDirectory}{The file path for the directory to download and store data in}

\item{lat}{The latitude for the MODIS in decimal degrees (not necessary if you do not need to download data)}

\item{long}{The longitude for the MODIS in decimal degrees (not necessary if you do not need to download data)}

\item{siteName}{The site name for file naming purposes}
}
\description{
Function to prepare MODIS data for input into forecast model
}
