% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalkPhenoModel2020.R
\name{randomWalkPhenoModel2020}
\alias{randomWalkPhenoModel2020}
\title{Creates a random walk phenology forecast model based on PhenoCam and MODIS data}
\usage{
randomWalkPhenoModel2020(data, nchain, priorCal = FALSE)
}
\arguments{
\item{data}{The data in the form of a list with data$p, data$mn, data$me, and data$n}

\item{nchain}{The desired number of chains in the MCMC}
}
\description{
Creates a random walk phenology forecast model based on PhenoCam and MODIS data
}
