% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleObs.R
\name{rescaleObs}
\alias{rescaleObs}
\title{Function to rescale phenology measurements to percent canopy cover for that calendar year}
\usage{
rescaleObs(time, vals)
}
\arguments{
\item{time}{A vector of days (x-values)}

\item{vals}{A vector of the measurements corresponding for the days (y-values)}
}
\description{
Function to rescale phenology measurements to percent canopy cover for that calendar year
}
