% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runForecastIter.R
\name{runForecastIter}
\alias{runForecastIter}
\title{Adds additional iterations until convergence and large enough effective sample size are met}
\usage{
runForecastIter(
  j.model,
  variableNames,
  maxIter = 10^9,
  baseNum = 5000,
  iterSize = 5000,
  ID = "",
  effSize = 5000,
  partialFile = FALSE,
  startFileName = FALSE
)
}
\arguments{
\item{j.model}{The JAGS model}

\item{variableNames}{Variable names for coda.samples}

\item{maxIter}{The maximum number of iterations to run before declaring that the model will not converge (optional)}

\item{baseNum}{The initial number of iterations to run before checking for convergence}

\item{iterSize}{The number of iterations to iteratively increase before checking for convergence again and large enough sample size (optional)}

\item{ID}{A identification string to paste with the message of non-convergence (optional)}

\item{effSize}{The desired necessary effective sample size (default is 5000 samples)}

\item{partialFile}{The desired name to save partial output (non-converged) to}

\item{startFileName}{The file name of a previous jags output that you want to add to (used to combine new runs with previous samples) (Note that this variable must be "partialOutput")}
}
\description{
Adds additional iterations until convergence and large enough effective sample size are met
}
