% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SW1PhenoModel.R
\name{SW1PhenoModel}
\alias{SW1PhenoModel}
\title{Creates the first spring warming model from Melaas 2016 GCB}
\usage{
SW1PhenoModel(data, nchain)
}
\arguments{
\item{data}{The data in the form of a list with data$p, data$mn, data$me, data$n, data$x_ic, and data$tau_ic}

\item{nchain}{The desired number of chains in the MCMC}
}
\description{
Creates the first spring warming model from Melaas 2016 GCB
}
