% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSf.R
\name{calSf}
\alias{calSf}
\title{Calculates cummulative Tair}
\usage{
calSf(Tairs, dates)
}
\arguments{
\item{Tairs}{The air temperature values in a vector}

\item{dates}{The desired dates}
}
\description{
Calculates cummulative Tair
}
