% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyPrevOutBurn.R
\name{identifyPrevOutBurn}
\alias{identifyPrevOutBurn}
\title{Identifies the most recent previous forecast file}
\usage{
identifyPrevOutBurn(siteName, forecastStr, startDate, endDate,
  dataDirectory)
}
\arguments{
\item{dataDirectory}{}
}
\description{
Identifies the most recent previous forecast file
}
