% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ERA5_Tair_New.R
\name{load_ERA5_Tair_New}
\alias{load_ERA5_Tair_New}
\title{Calculates cummulative Tair (within one year for one ensemble)}
\usage{
load_ERA5_Tair_New(
  lat = "",
  long = "",
  endDate = "",
  calDatesT = TRUE,
  ERA5dataFolder,
  TZ_name = "America/New_York"
)
}
\arguments{
\item{lat}{The site latitude}

\item{long}{The site longitude}
}
\description{
Calculates cummulative Tair (within one year for one ensemble)
}
