% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GEFS_Forecast.R
\name{load_GEFS_Forecast}
\alias{load_GEFS_Forecast}
\title{Loads the GEFS weather forecast data}
\usage{
load_GEFS_Forecast(fileName)
}
\arguments{
\item{fileName}{The GEFS filename}
}
\description{
Loads the GEFS weather forecast data
}
