% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_NOAA_met.R
\name{load_NOAA_met}
\alias{load_NOAA_met}
\title{Loads NOAA met data}
\usage{
load_NOAA_met(station, startDate = "", endDate = "")
}
\arguments{
\item{station}{The name of the NOAA station}
}
\description{
Loads NOAA met data
}
