% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticPhenoModel.R
\name{logisticPhenoModel}
\alias{logisticPhenoModel}
\title{Creates a logistic phenology forecast model based on PhenoCam and MODIS data}
\usage{
logisticPhenoModel(data, nchain)
}
\arguments{
\item{data}{The data in the form of a list with data$p, data$mn, data$me, data$n, data$x_ic, and data$tau_ic}

\item{nchain}{The desired number of chains in the MCMC}
}
\description{
Creates a logistic phenology forecast model based on PhenoCam and MODIS data
}
