% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyForecast.R
\name{phenologyForecast}
\alias{phenologyForecast}
\title{General phenology forecast function}
\usage{
phenologyForecast(forecastType, forecastLength = 16, siteName, URL, lat, long,
  dataDirectory, startDate, endDate, cValsPC, dValsPC, cValsMN, dValsMN,
  cValsME, dValsME, GEFS_Files = "", GEFS_Directory)
}
\arguments{
\item{forecastType}{The type of forecast (randomWalk or logistic)}

\item{forecastLength}{The length of the forecast into the future in days}

\item{siteName}{The site name}

\item{URL}{The URL where the site's phenocam data is located}

\item{lat}{The latitude of the site in decimals}

\item{long}{The longitude of the site in decimals}

\item{dataDirectory}{The file path for the directory to download and store data in}

\item{startDate}{The start date for the forecast in date form}

\item{endDate}{The end date for the forecast in date form}

\item{cValsPC}{The c values for rescaling for PhenoCam}

\item{dValsPC}{The d values for rescaling for PhenoCam}

\item{cValsMN}{The c values for rescaling for MODIS NDVI}

\item{dValsMN}{The d values for rescaling for MODIS NDVI}

\item{cValsME}{The c values for rescaling for MODIS EVI}

\item{dValsME}{The d values for rescaling for MODIS EVI}

\item{GEFS_Files}{The filenames for the GEFS files}

\item{GEFS_Directory}{The directory where the GEFS files are located}
}
\description{
General phenology forecast function
}
