% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyForecast_Autumn.R,
%   R/phenologyForecast_Autumn2020.R, R/phenologyForecast_Spring.R
\name{phenologyForecast_Autumn}
\alias{phenologyForecast_Autumn}
\title{General phenology forecast function}
\usage{
phenologyForecast_Autumn(
  forecastType,
  forecastLength = 14,
  siteName,
  URLs,
  lat,
  long,
  dataDirectory,
  startDate,
  endDate,
  cValsPC,
  dValsPC,
  cValsMN,
  dValsMN,
  cValsME,
  dValsME,
  GEFS_Files = "",
  GEFS_Directory,
  station = "",
  season,
  baseTemp = NA,
  index = "all",
  prevOutBurn = NA,
  baseNum = 10000,
  iterSize = 5000,
  effSize = 5000,
  partialFile = FALSE
)

phenologyForecast_Autumn(
  forecastType,
  forecastLength = 14,
  siteName,
  URLs,
  lat,
  long,
  dataDirectory,
  startDate,
  endDate,
  cValsPC,
  dValsPC,
  cValsMN,
  dValsMN,
  cValsME,
  dValsME,
  GEFS_Files = "",
  GEFS_Directory,
  station = "",
  season,
  baseTemp = NA,
  index = "all",
  prevOutBurn = NA,
  baseNum = 10000,
  iterSize = 5000,
  effSize = 5000,
  partialFile = FALSE
)

phenologyForecast_Autumn(
  forecastType,
  forecastLength = 14,
  siteName,
  URLs,
  lat,
  long,
  dataDirectory,
  startDate,
  endDate,
  cValsPC,
  dValsPC,
  cValsMN,
  dValsMN,
  cValsME,
  dValsME,
  GEFS_Files = "",
  GEFS_Directory,
  station = "",
  season,
  baseTemp = NA,
  index = "all",
  prevOutBurn = NA,
  baseNum = 10000,
  iterSize = 5000,
  effSize = 5000,
  partialFile = FALSE
)
}
\arguments{
\item{forecastType}{The type of forecast (randomWalk or logistic)}

\item{forecastLength}{The length of the forecast into the future in days}

\item{siteName}{The site name}

\item{URLs}{The URL where the site's phenocam data is located}

\item{lat}{The latitude of the site in decimals}

\item{long}{The longitude of the site in decimals}

\item{dataDirectory}{The file path for the directory to download and store data in}

\item{startDate}{The start date for the forecast in date form}

\item{endDate}{The end date for the forecast in date form}

\item{cValsPC}{The c values for rescaling for PhenoCam}

\item{dValsPC}{The d values for rescaling for PhenoCam}

\item{cValsMN}{The c values for rescaling for MODIS NDVI}

\item{dValsMN}{The d values for rescaling for MODIS NDVI}

\item{cValsME}{The c values for rescaling for MODIS EVI}

\item{dValsME}{The d values for rescaling for MODIS EVI}

\item{GEFS_Files}{The filenames for the GEFS files}

\item{GEFS_Directory}{The directory where the GEFS files are located}

\item{season}{The desired season: spring or fall}

\item{baseTemp}{If CDD, the desired base temperature}

\item{index}{The desired index (GCC, NDVI, EVI, or all)}

\item{prevOutBurn}{The most recent previous forecast output}
}
\description{
General phenology forecast function

General phenology forecast function

General phenology forecast function
}
