% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleObs.R
\name{rescaleObs}
\alias{rescaleObs}
\title{Function to rescale phenology measurements to percent canopy cover for that calendar year}
\usage{
rescaleObs(times, vals, cVals, dVals, partialStart = FALSE)
}
\arguments{
\item{times}{A vector of days (x-values)}

\item{vals}{A vector of the measurements corresponding for the days (y-values)}

\item{cVals}{The c values for rescaling}

\item{dVals}{The d values for rescaling}

\item{partialStart}{Whether or not there is a partial year of data included (e.g. with random walk)}
}
\description{
Function to rescale phenology measurements to percent canopy cover for that calendar year
}
